% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.model.output.dalec.R
\name{get.model.output.dalec}
\alias{get.model.output.dalec}
\title{Retrieve model output from local or remote server}
\usage{
get.model.output.dalec()
}
\description{
Retrieve model output from local or remote server
}
