% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.DALEC.R
\name{met2model.DALEC}
\alias{met2model.DALEC}
\title{met2model.DALEC}
\usage{
met2model.DALEC(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  spin_nyear = NULL,
  spin_nsample = NULL,
  spin_resample = NULL,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{spin_nyear, spin_nsample, spin_resample}{passed on to
`PEcAn.data.atmosphere::spin.met()`}

\item{...}{additional arguments, currently ignored}
}
\value{
Invisibly returns a data.frame with metadata for the generated DALEC meteorology file.
}
\description{
met2model for DALEC
}
\examples{
\dontrun{
 in.path    <- "~/paleon/PalEONregional_CF_site_1-24047/"
 in.prefix  <- ""
 outfolder  <- "~/paleon/metTest/"
 merge.file <- "~/paleon/paleon_monthly_co2.nc"
 start_date <- "0850-01-01"
 end_date   <- "2010-12-31"
 PEcAn.DALEC::met2model.DALEC(in.path, in.prefix, outfolder, start_date, end_date)
}
}
