% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model2netcdf.DALEC.R
\name{model2netcdf.DALEC}
\alias{model2netcdf.DALEC}
\title{Function to convert DALEC model output to standard netCDF format}
\usage{
model2netcdf.DALEC(outdir, sitelat, sitelon, start_date, end_date)
}
\arguments{
\item{outdir}{Location of DALEC model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert DALEC output to netCDF
}
\details{
Converts all output contained in a folder to netCDF.
}
\author{
Shawn Serbin, Michael Dietze
}

