% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.config.DALEC}
\alias{write.config.DALEC}
\title{write.config.DALEC}
\usage{
write.config.DALEC(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{}

\item{trait.values}{}

\item{settings}{}

\item{run.id}{}
}
\value{
configuration files
}
\description{
Writes a configuration files for your model
write Dalec Configuration files
}

