% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.configs.dalec.R
\name{write.config.DALEC}
\alias{write.config.DALEC}
\title{write.config.DALEC}
\usage{
write.config.DALEC(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{ignored}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{a PEcAn settings object}

\item{run.id}{Unique identifier for the run,
used here to construct output directories and filenames.}
}
\value{
configuration files
}
\description{
Writes a configuration files for your model
write Dalec Configuration files
}
