% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.configs.dalec.R
\name{write.config.DALEC}
\alias{write.config.DALEC}
\title{Writes a configuration files for your model}
\usage{
write.config.DALEC(defaults, trait.values, settings, run.id)
}
\description{
Writes a configuration files for your model
}

